Reduce dependency on instance variables
We've been caught by this once because it's harder to read -- by passing
our credentials directly to our client helper - and receiving client in
return, rather than populating an instance variable -- this code should
be a little easier to read and safer to use.

Note that this uses the long-form list() destructuring support. As of
PHP 7.1, there's also shorthand destructuring:

```php
list($a, $b) = $arr;
// becomes
[$a, $b] = $arr;
```

However, we still support Magento 2.2 and Magento 2.2 still supports PHP
7.0.